Added category service exceptions

+++ src/main/java/com/bsuir/petition/service/category/impl/CategoryServiceImpl.java
@@ -6,11 +6,13 @@ import com.bsuir.petition.bean.entity.Category;
 import com.bsuir.petition.dao.CategoryDao;
 import com.bsuir.petition.service.category.CategoryService;
 import com.bsuir.petition.service.category.exception.CategoryNotFoundException;
+import com.bsuir.petition.service.category.exception.SuchCategoryExistsException;
 import com.bsuir.petition.service.category.util.CategoryDataValidator;
 import com.bsuir.petition.service.category.util.DtoExchangerCategory;
 import com.bsuir.petition.service.category.util.ExchangerCategory;
 import com.bsuir.petition.service.exception.ErrorInputException;
 import com.bsuir.petition.service.exception.ServerException;
+import org.hibernate.HibernateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -51,8 +53,12 @@ public class CategoryServiceImpl implements CategoryService {
     public CategoryListDTO getCategories() throws ServerException {
         CategoryListDTO categoryListDTO;
         List<Category> categories;
-        categories = categoryDao.getCategories();
-        categoryListDTO = dtoExchangerCategory.getCategoryListDTO(categories);
+        try {
+            categories = categoryDao.getCategories();
+            categoryListDTO = dtoExchangerCategory.getCategoryListDTO(categories);
+        } catch (HibernateException exception) {
+            throw new ServerException("Server exception!", exception);
+        }
         return categoryListDTO;
     }
 
@@ -60,7 +66,16 @@ public class CategoryServiceImpl implements CategoryService {
     public void deleteCategory(long id) throws CategoryNotFoundException, ServerException {
         Category category;
         category = categoryDao.getCategory(id);
-        categoryDao.deleteCategory(category);
+
+        if (category == null) {
+            throw new CategoryNotFoundException("Category not found!");
+        }
+
+        try {
+            categoryDao.deleteCategory(category);
+        } catch (HibernateException exception) {
+            throw new ServerException("Such category does", exception);
+        }
     }
 
     @Override
@@ -70,16 +85,24 @@ public class CategoryServiceImpl implements CategoryService {
 
         Category category;
         category = exchangerCategory.getCategory(shortCategoryDTO, id);
-        categoryDao.updateCategory(category);
+        try {
+            categoryDao.updateCategory(category);
+        } catch (HibernateException exception) {
+            throw new CategoryNotFoundException("No such category!", exception);
+        }
     }
 
     @Override
-    public void addCategory(ShortCategoryDTO shortCategoryDTO) throws ErrorInputException, ServerException {
+    public void addCategory(ShortCategoryDTO shortCategoryDTO) throws ErrorInputException, ServerException, SuchCategoryExistsException {
 
         categoryDataValidator.validate(shortCategoryDTO);
 
         Category category;
         category = exchangerCategory.getCategory(shortCategoryDTO);
-        categoryDao.addCategory(category);
+        try {
+            categoryDao.addCategory(category);
+        } catch (HibernateException exception) {
+            throw new SuchCategoryExistsException("Server exception!", exception);
+        }
     }
 }
